BIG refactor - split class into files for easier navigation. Tabs converted into 4 spaces to preserve compatibility
